Rename qt_layers_list to qt_layerlist
