+++ napari/_qt/qt_layerlist.py
